requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-nmap>=0.7.1
scapy>=2.5.0
netifaces>=0.11.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
mac-vendor-lookup>=0.1.12

# Network Utilities
ipaddress>=1.0.23
requests>=2.31.0

# Data Processing
pandas>=2.1.3
numpy>=1.26.2

# Process Management
psutil>=5.9.6

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# Date/Time
python-dateutil>=2.8.2

# Async Support
asyncio-throttle>=1.0.2

# Type Hints
typing-extensions>=4.8.0
